+++ packages/theme/src/node/compact/navbar.ts
@@ -1,30 +1,37 @@
 import { isArray, isPlainObject, isString } from "@vuepress/shared";
+import { colors } from "@vuepress/utils";
 import { deprecatedLogger } from "./utils.js";
 import { logger } from "../utils.js";
 
 import type {
+  AutoLinkOptions,
   NavbarGroup,
   NavbarItem,
   NavbarOptions,
 } from "../../shared/index.js";
 
-const handleNavbarOptions = (config: unknown[]): NavbarOptions =>
+type LegacyNavbarOptions = (
+  | string
+  | AutoLinkOptions
+  | (NavbarGroup & { items?: (AutoLinkOptions | string)[] })
+)[];
+
+const handleNavbarOptions = (config: LegacyNavbarOptions): NavbarOptions =>
   config
     .map((item) => {
       if (isString(item)) return item;
 
       if (isPlainObject(item) && item) {
         deprecatedLogger({
-          options: item as Record<string, unknown>,
+          // @ts-ignore
+          options: item,
           deprecatedOption: "items",
           newOption: "children",
           scope: "navbar",
         });
 
-        // @ts-ignore
-        if (isArray(item.children))
-          // @ts-ignore
-          handleNavbarOptions(item.children as unknown[]);
+        if ("children" in item && isArray(item.children))
+          handleNavbarOptions(item.children);
 
         return item as NavbarItem | NavbarGroup;
       }
@@ -37,12 +44,12 @@ const handleNavbarOptions = (config: unknown[]): NavbarOptions =>
  * @deprecated You should use V2 standard navbar config and avoid using it
  */
 export const convertNavbarOptions = (
-  config: unknown
+  config: NavbarOptions | unknown
 ): NavbarOptions | false => {
   if (config === false) return false;
-  if (isArray(config)) return handleNavbarOptions(config);
+  if (isArray(config)) return handleNavbarOptions(config as NavbarOptions);
 
-  logger.error('"navbar" config should be an array');
+  logger.error(`${colors.magenta("navbar")} config should be an array`);
 
   return false;
 };