+++ packages/theme/src/node/compact/sidebar.ts
@@ -1,4 +1,5 @@
 import { isArray, isPlainObject, isString } from "@vuepress/shared";
+import { colors } from "@vuepress/utils";
 import { deprecatedLogger, droppedLogger } from "./utils.js";
 import { logger } from "../utils.js";
 
@@ -8,7 +9,9 @@ import type {
   SidebarOptions,
 } from "../../shared/index.js";
 
-const handleArraySidebarOptions = (config: unknown[]): SidebarArrayOptions =>
+const handleArraySidebarOptions = (
+  config: SidebarArrayOptions
+): SidebarArrayOptions =>
   config
     .map((item) => {
       if (isString(item)) return item;
@@ -22,23 +25,19 @@ const handleArraySidebarOptions = (config: unknown[]): SidebarArrayOptions =>
 
         convertConfig.forEach(([deprecatedOption, newOption]) =>
           deprecatedLogger({
-            options: item as Record<string, unknown>,
+            // @ts-ignore
+            options: item,
             deprecatedOption,
             newOption,
             scope: "sidebar",
           })
         );
 
-        droppedLogger(
-          item as Record<string, unknown>,
-          "sidebarDepth",
-          "Found in sidebar"
-        );
-
         // @ts-ignore
-        if (isArray(item.children))
-          // @ts-ignore
-          handleArraySidebarOptions(item.children as unknown[]);
+        droppedLogger(item, "sidebarDepth", "Found in sidebar");
+
+        if ("children" in item && isArray(item.children))
+          handleArraySidebarOptions(item.children);
 
         return item as SidebarItem;
       }
@@ -51,30 +50,37 @@ const handleArraySidebarOptions = (config: unknown[]): SidebarArrayOptions =>
  * @deprecated You should use V2 standard sidebar config and avoid using it
  */
 export const convertSidebarOptions = (
-  config: unknown
+  config: SidebarOptions | unknown
 ): SidebarOptions | false => {
-  if (config === false) return false;
-  if (isArray(config)) return handleArraySidebarOptions(config);
+  if (config === false || config === "structure") return config;
+
+  if (isArray(config))
+    return handleArraySidebarOptions(config as SidebarArrayOptions);
 
   if (isPlainObject(config) && config)
     return Object.fromEntries(
       Object.entries(config).map<
-        [string, SidebarArrayOptions | "structure" | false]
+        [string, SidebarArrayOptions | "structure" | "heading" | false]
       >(([key, value]) => {
-        if (isArray(value)) return [key, handleArraySidebarOptions(value)];
+        if (isArray(value))
+          return [key, handleArraySidebarOptions(value as SidebarArrayOptions)];
 
-        if (value === "structure" || value === false)
-          return [key, <"structure" | false>value];
+        if (value === "structure" || value === "heading" || value === false)
+          return [key, <"structure" | "heading" | false>value];
 
         logger.error(
-          '"sidebar" value should be an array, "structure" or false when setting as an object'
+          '"sidebar" value should be an array, "structure", "heading" or false when setting as an object'
         );
 
         return [key, false];
       })
     );
 
-  logger.error('"sidebar" config should be an array or object');
+  logger.error(
+    `${colors.magenta(
+      "sidebar"
+    )} config should be: an array, an object, "structure", "heading" or false`
+  );
 
   return false;
 };