+++ packages/theme/src/shared/options/layout/sidebar.ts
@@ -64,7 +64,7 @@ export interface SidebarLocaleOptions {
    * @description 设置为 `false` 以在当前语言中禁用侧边栏
    * @see https://theme-hope.vuejs.vuepress/zh/guide/layout/sidebar.html
    */
-  sidebar?: SidebarOptions | "structure" | "heading" | false;
+  sidebar?: SidebarOptions;
 
   /**
    * Whether show icons in the sidebar